Extract gradlebuild.module-jar plugin
for Gradle module jar archive name and classpath manifest generation
