+++ build-logic-commons/gradle-plugin/src/main/kotlin/gradlebuild.kotlin-shared-runtime.gradle.kts
@@ -1,10 +1,9 @@
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
     id("java-library")
     kotlin("jvm")
-    id("gradlebuild.module-identity")
+    id("gradlebuild.module-jar")
     id("gradlebuild.reproducible-archives")
     id("gradlebuild.repositories")
     id("gradlebuild.code-quality")
